Added documentation to Wrapper.cs methods, added Include and Exclude parameters
Each Get method now has include and exclude. These are string[]. At least one of them (or both of them) must be NULL.

Include array specifies fields that will be returned in response. All other fields will be NULL (or default value). If this include array is null, whole reponse is returned.

Exclude array specifies fields that will NOT be returned in response, so they will be NULL (or default value). Other fields will be returned as usuall. If this exclude array is null, whole reponse is returned.