fix null reference exception on cache.clear
This change checks if the "cache" is already empty before clearing it, to avoid a null reference exception. You may want to go further in depth as to why it's null at this point if you don't expect it to be, but this at least gets the API working again for me.