Updated to Version 0.7.0
This package was updated to version 0.7.0. This update changed platform that was library developed on, to .Net Standart.

In this commit, new folder "CR API Standart Wrapper" was created and all files were copied there. The new folder contains the .Net Standart project. In next update, the old folder will be removed.